pkg/k8s: refactor constants to a dedicated package
This will allow to import constant names from other packages without
worrying on importing k8s dependencies.
