+++ pkg/k8s/constants/const.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2017 Authors of Cilium
+// Copyright 2016-2020 Authors of Cilium
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,17 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8s
-
-import (
-	"time"
-)
+package constants
 
 const (
-	// BackOffLoopTimeout is the default duration when trying to reach the
-	// kube-apiserver.
-	BackOffLoopTimeout = 2 * time.Minute
-
 	// EnvNodeNameSpec is the environment label used by Kubernetes to
 	// specify the node's name.
 	EnvNodeNameSpec = "K8S_NODE_NAME"
